+++ app/controllers/recording_studio_presskits/application_controller.rb
@@ -13,6 +13,7 @@ module RecordingStudioPresskits
 
     helper ::RecordingStudio::LayoutHelper if defined?(::RecordingStudio::LayoutHelper)
     helper ::RecordingStudioAccessible::AvatarsHelper if defined?(::RecordingStudioAccessible::AvatarsHelper)
+    helper ::RecordingStudioPublishable::Engine.helpers if defined?(::RecordingStudioPublishable::Engine)
 
     before_action :authenticate_presskits_actor!
     before_action :set_presskits_current_actor